[DataLoader] DataLoader now automatically apply sharding to DataPipes (#78762)
ghstack-source-id: ac918b064cd09cd68a04c28238481c76b46b4010
Pull Request resolved: https://github.com/pytorch/pytorch/pull/78631
